feat: 채팅방 조회 기능 서비스로직 구현

+++ src/main/kotlin/com/example/backend/domain/chatroom/service/ChatRoomService.kt
@@ -4,6 +4,7 @@ import com.example.backend.domain.chatroom.dto.request.CreateChatRoomRequest
 import com.example.backend.domain.chatroom.dto.response.ChatRoomResponse
 import com.example.backend.domain.chatroom.entity.ChatRoom
 import com.example.backend.domain.chatroom.entity.enums.ChatRoomType
+import org.springframework.transaction.annotation.Transactional
 import com.example.backend.domain.chatroom.exception.ChatRoomException
 import com.example.backend.domain.chatroom.exception.message.ChatRoomExceptionMessage
 import com.example.backend.domain.chatroom.repository.ChatRoomRepository
@@ -11,7 +12,6 @@ import com.example.backend.domain.user.entity.User
 import com.example.backend.domain.user.exception.UserException
 import com.example.backend.domain.user.exception.message.UserExceptionMessage
 import com.example.backend.domain.user.repository.UserRepository
-import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 
 @Service
@@ -19,6 +19,21 @@ class ChatRoomService (
     private val chatRoomRepository: ChatRoomRepository,
     private val userRepository: UserRepository
 ) {
+    // 채팅방 전체 조회
+    @Transactional(readOnly = true)
+    fun getAllChatRooms(): List<ChatRoomResponse> {
+        val rooms = chatRoomRepository.findAll()
+        return if (rooms.isEmpty()) emptyList() else rooms.map { it.toResponse() }
+    }
+
+    // 채팅방 상세 조회
+    @Transactional(readOnly = true)
+    fun getChatRoom(chatRoomId: Long): ChatRoomResponse {
+        val chatRoom = chatRoomRepository.findById(chatRoomId)
+            .orElseThrow { ChatRoomException(ChatRoomExceptionMessage.NOT_FOUND_CHATROOM) }
+        return chatRoom.toResponse()
+    }
+
     //채팅방 생성
     @Transactional
     fun createChatRoom(
@@ -31,6 +46,7 @@ class ChatRoomService (
         val maxMembers = if (request.chatRoomType == ChatRoomType.PERSONAL) 2 else request.maxMembers
 
         val chatRoom = ChatRoom(
+            creatorId = creator.id,
             chatRoomName = request.chatRoomName,
             description = request.description,
             maxMembers = maxMembers,
@@ -47,9 +63,11 @@ class ChatRoomService (
     @Transactional
     fun updateChatRoomName(
         chatRoomId: Long,
+        userId: Long,
         chatRoomName: String
     ) {
         val chatRoom = findChatRoom(chatRoomId)
+        checkPermission(chatRoomId, userId)
         chatRoom.updateName(chatRoomName)
 
     }
@@ -58,16 +76,21 @@ class ChatRoomService (
     @Transactional
     fun updateChatRoomDescription(
         chatRoomId: Long,
+        userId: Long,
         chatRoomDescription: String
     ) {
         val chatRoom = findChatRoom(chatRoomId)
+        checkPermission(chatRoomId, userId)
         chatRoom.updateDescription(chatRoomDescription)
     }
 
     //채팅방 삭제
     @Transactional
-    fun deleteChatRoom(chatRoomId: Long) {
+    fun deleteChatRoom(
+        chatRoomId: Long,
+        userId: Long) {
         val chatRoom = findChatRoom(chatRoomId)
+        checkPermission(chatRoomId, userId)
         chatRoomRepository.delete(chatRoom)
     }
 
@@ -75,8 +98,7 @@ class ChatRoomService (
     @Transactional
     fun joinChatRoom(chatRoomId: Long, userId: Long) {
         val chatRoom = findChatRoom(chatRoomId)
-        val user = userRepository.findById(userId)
-            .orElseThrow { UserException(UserExceptionMessage.NOT_FOUND_USER) }
+        val user = findUser(userId)
         chatRoom.addUser(user);
     }
 
@@ -84,14 +106,29 @@ class ChatRoomService (
     @Transactional
     fun leaveChatRoom(chatRoomId: Long,userId: Long) {
         val chatRoom = findChatRoom(chatRoomId)
-        val user = userRepository.findById(userId)
-            .orElseThrow { UserException(UserExceptionMessage.NOT_FOUND_USER) }
+        val user = findUser(userId)
+        if (userId == chatRoom.creatorId) {
+            throw ChatRoomException(ChatRoomExceptionMessage.REQUIRED_CHATROOM_DESCRIPTION)
+        }
         chatRoom.removeUser(user);
     }
 
     fun findChatRoom(chatRoomId: Long): ChatRoom =
         chatRoomRepository.findById(chatRoomId)
             .orElseThrow { ChatRoomException(ChatRoomExceptionMessage.NOT_FOUND_CHATROOM) }
+
+    fun findUser(userId: Long): User =
+        userRepository.findById(userId)
+            .orElseThrow{ UserException(UserExceptionMessage.NOT_FOUND_USER) }
+
+    fun checkPermission(chatRoomId: Long, userId: Long): Boolean {
+        val user = findUser(userId)
+        val chatRoom = findChatRoom(chatRoomId)
+        if(user.id != chatRoom.creatorId){
+            throw ChatRoomException(ChatRoomExceptionMessage.NO_PERMISSION)
+        }
+        return true
+    }
 
     fun ChatRoom.toResponse(): ChatRoomResponse =
         ChatRoomResponse(
@@ -102,6 +139,6 @@ class ChatRoomService (
             chatRoomType = this.chatRoomType,
             participation = this.participation,
             userCount = this.users.size,
-            creatorId = this.users.minBy { it.id!! }.id!!
+            creatorId = this.creatorId
         )
 }
